p added

+++ include/function.h
@@ -1,7 +1,7 @@
 #ifndef FUNCTION_
 #define FUNCTION_
 
-int print_base(long long nb, char *base, int fd);
+int print_base(long nb, char *base, int fd);
 int stu_dprintf(int fd, const char *pattern, ...);
 
 #endif
